Налажена отрисовка дней, изменение общей стоимости и Инфо о маршруте при изменении точек маршрута
