+++ src/controllers/point.js
@@ -1,9 +1,6 @@
 import TripCardComponent from "../components/trip-card";
 import FormEditComponent from "../components/form-edit";
 import {render, RenderPosition, remove, replace} from "../utils/render";
-// import {getRandomElement} from "../utils/common";
-// import {cities, eventTypes} from "../const";
-// import {generateDescriptionText, generateOptionsList} from "../mock/trip-card";
 
 export const Mode = {
   ADDING: `adding`,
@@ -114,7 +111,6 @@ export default class PointController {
   }
 
   destroy() {
-    // console.log(this._editCardComponent);
     remove(this._editCardComponent);
     remove(this._cardComponent);
     document.removeEventListener(`keydown`, this._onEscKeyDown);
